Added redacted.rb file
This is a small ruby program which will take a userInput and a list of redacted words (separated by comma). It will find and replace with asterisks any word that needs to be redacted. The program is able to find words that match even if they do not match the case.